requirements.txt
# Web framework
Flask==2.3.3

# Environment variable handling
python-dotenv==1.0.1

# Computer vision
opencv-python==4.8.1.78
numpy==1.26.4

# YOLOv8 Ultralytics
ultralytics==8.0.134

# Google Generative AI (Gemini)
google-generativeai==0.6.3

# Optional: JSON handling (built-in Python, no need to install)
# json (built-in)

# Other utilities (optional but recommended)
requests==2.32.1
